Add reminder scheduler; enchance sleep page & small fix

+++ Android/app/src/main/java/mad/project/mdp_project/data/SleepLog.kt
@@ -11,7 +11,7 @@ data class SleepLog(
     val userId: Int,
     val startTime: Long,
     val endTime: Long,
-    val quality: Int,
+    val quality: Float,
     val date: Long = System.currentTimeMillis()
 ) {
     fun getDurationMillis(): Long = endTime - startTime
